moved sys_mempool struct to sys_mempool.h
